tests: merge two spec files
